+++ jasperreports/src/net/sf/jasperreports/engine/fill/JRFillChart.java
@@ -1384,7 +1384,7 @@ public class JRFillChart extends JRFillElement implements JRChart
 
 		Color color = interval.getBackgroundColor();
 		float[] components = color.getRGBColorComponents(null);
-		float alpha = interval.getAlphaDouble() == null ? (float)JRMeterInterval.DEFAULT_TRANSPARENCY : interval.getAlphaDouble().floatValue();
+		float alpha = (float)(interval.getAlphaDouble() == null ? JRMeterInterval.DEFAULT_TRANSPARENCY : interval.getAlphaDouble());
 
 		Color alphaColor = new Color(components[0], components[1], components[2], alpha);
